+++ mem/storager.go
@@ -1,12 +1,14 @@
 package mem
 
 import (
+	"github.com/viant/afs/base"
 	"github.com/viant/afs/file"
 	"github.com/viant/afs/storage"
 	"github.com/viant/afs/url"
 )
 
 type storager struct {
+	base.Storager
 	scheme string
 	Root   *Folder
 }
@@ -17,8 +19,10 @@ func (s *storager) Close() error {
 
 //NewStorager create a new in memeory storage service
 func NewStorager(baseURL string) storage.Storager {
-	return &storager{
+	result := &storager{
 		Root:   NewFolder(baseURL, file.DefaultDirOsMode),
 		scheme: url.Scheme(baseURL, Scheme),
 	}
+	result.Storager.List = result.List
+	return result
 }